+++ client/src/routes/QuestionForm.js
@@ -1,0 +1,9 @@
+import StartSection from "../components/StartSection";
+
+export const QuestionForm = () => {
+  return (
+    <>
+      <StartSection />
+    </>
+  );
+};
