+++ client/src/styles.css
@@ -1,27 +1,20 @@
-
 :root {
   --font: "EB Garamond", serif;
-	--font-title: "Playfair Display", serif;
-	--font-sans-serif: "Manrope", sans-serif;
-	--black: #1c1f33;
-	--black--acc: #222;
-	--gray: #eee;
-	--gray-1: #ededed;
+  --font-title: "Playfair Display", serif;
+  --font-sans-serif: "Manrope", sans-serif;
+  --black: #1c1f33;
+  --black--acc: #222;
+  --gray: #eee;
+  --gray-1: #ededed;
   color: white;
 }
 
 :root,
 body {
+  background-color: black;
   margin: 0;
   padding: 0;
 }
-
-
-.background{
-  background-color: black;
-  z-index: -1;
-}
-
 
 h1 {
   font-size: 5rem;
@@ -29,7 +22,7 @@ h1 {
   text-align: center;
 }
 
-h3{
+h3 {
   text-align: center;
 }
 
@@ -43,7 +36,7 @@ button {
   font-size: 1rem;
 }
 
-.btn-group{
+.btn-group {
   display: flex;
   margin-top: 20px;
 }
@@ -72,7 +65,7 @@ button {
   font-family: serif;
   color: #333;
   font-size: 1.2rem;
-	margin-top: 10px;
+  margin-top: 10px;
   padding: 1.5rem 2rem;
   border-radius: 0.2rem;
   background-color: rgb(255, 255, 255);
